use rhsa template for CVE flaws
